+++ fstat_na.go
@@ -15,5 +15,5 @@ func fstat(f *os.File) (FileInfo, error) {
 		return nil, &os.PathError{Op: "stat", Path: "", Err: os.ErrInvalid}
 	}
 
-	return nil, &os.PathError{Op: "stat", Path: f.Name(), Err: errors.ErrUnsupported}
+	return nil, &os.PathError{Op: "stat", Path: f.Name(), Err: &NotYetImplementedError("fstat")}
 }