+++ rename_plan9.go
@@ -40,7 +40,7 @@ func rename(source, destination string, opts ...Option) error {
 			return renameError(
 				source,
 				destination,
-				errors.ErrUnsupported,
+				&compat.UnsupportedError{"rename"},
 			)
 		}