+++ volume/volume_na.go
@@ -1,12 +1,14 @@
 // SPDX-FileCopyrightText: Copyright © 2025 Ross Smith II <ross@smithii.com>
 // SPDX-License-Identifier: MIT
 
-//go:build plan9
+//go:build (!unix && !windows) || (openbsd && ppc64) || (netbsd && 386) || (freebsd && riscv64) || (cgo && aix && ppc64)
 
 package volume
 
 import (
 	"errors"
+
+	"github.com/rasa/compat"
 )
 
 var osFeatureMap = map[OSFeature]string{}
@@ -14,5 +16,5 @@ var osFeatureMap = map[OSFeature]string{}
 func Volumes(mounts []Mount) ([]Volume, error) {
 	volumes := []Volume{}
 
-	return volumes, errors.ErrUnsupported
+	return volumes, &compat.NotYetImplementedError("volumes")
 }